[FLINK-11282][network] Merge StreamRecordWriter into RecordWriter
