+++ flink-streaming-java/src/test/java/org/apache/flink/streaming/runtime/io/benchmark/LongRecordWriterThread.java
@@ -20,7 +20,6 @@ package org.apache.flink.streaming.runtime.io.benchmark;
 
 import org.apache.flink.core.testutils.CheckedThread;
 import org.apache.flink.runtime.io.network.api.writer.RecordWriter;
-import org.apache.flink.streaming.runtime.io.StreamRecordWriter;
 import org.apache.flink.types.LongValue;
 
 import java.io.IOException;
@@ -34,7 +33,7 @@ import static org.apache.flink.util.Preconditions.checkState;
  * records.
  */
 public class LongRecordWriterThread extends CheckedThread {
-	private final StreamRecordWriter<LongValue> recordWriter;
+	private final RecordWriter<LongValue> recordWriter;
 	private final boolean broadcastMode;
 
 	/**
@@ -45,7 +44,7 @@ public class LongRecordWriterThread extends CheckedThread {
 	private volatile boolean running = true;
 
 	public LongRecordWriterThread(
-			StreamRecordWriter<LongValue> recordWriter,
+			RecordWriter<LongValue> recordWriter,
 			boolean broadcastMode) {
 		this.recordWriter = checkNotNull(recordWriter);
 		this.broadcastMode = broadcastMode;
