+++ flink-streaming-java/src/test/java/org/apache/flink/streaming/runtime/io/benchmark/StreamNetworkPointToPointBenchmark.java
@@ -71,7 +71,7 @@ public class StreamNetworkPointToPointBenchmark {
 	 *
 	 * @param flushTimeout
 	 * 		output flushing interval of the
-	 * 		{@link org.apache.flink.streaming.runtime.io.StreamRecordWriter}'s output flusher thread
+	 * 		{@link org.apache.flink.runtime.io.network.api.writer}'s output flusher thread
 	 */
 	public void setUp(long flushTimeout, Configuration config) throws Exception {
 		environment = new StreamNetworkBenchmarkEnvironment<>();
